refactor: extracted barcode icon :recycle:
