+++ src/commons/components/Icons/index.tsx
@@ -1,5 +1,6 @@
 import AlertIconComponent from "./AlertIcon/AlertIcon.component";
 import ComputerIconComponent from "./ComputerIcon/ComputerIcon.component";
 import WheelIconComponent from "./WheelIcon/WheelIcon.component";
+import CodebarIconComponent from "./CodeBarIcon/CodeBarIcon.component";
 
-export {AlertIconComponent, ComputerIconComponent, WheelIconComponent};
+export {AlertIconComponent, ComputerIconComponent, WheelIconComponent, CodebarIconComponent};
